+++ component/processor_test.go
@@ -62,6 +62,31 @@ func TestNewProcessorFactory_WithOptions(t *testing.T) {
 	assert.NoError(t, err)
 }
 
+func TestNewProcessorFactory_WithStabilityLevel(t *testing.T) {
+	const typeStr = "test"
+	defaultCfg := config.NewProcessorSettings(config.NewComponentID(typeStr))
+	factory := NewProcessorFactory(
+		typeStr,
+		func() config.Processor { return &defaultCfg },
+		WithTracesProcessorAndStabilityLevel(createTracesProcessor, StabilityLevelAlpha),
+		WithMetricsProcessorAndStabilityLevel(createMetricsProcessor, StabilityLevelBeta),
+		WithLogsProcessorAndStabilityLevel(createLogsProcessor, StabilityLevelUnmaintained))
+	assert.EqualValues(t, typeStr, factory.Type())
+	assert.EqualValues(t, &defaultCfg, factory.CreateDefaultConfig())
+
+	assert.EqualValues(t, StabilityLevelAlpha, factory.StabilityLevel(config.TracesDataType))
+	_, err := factory.CreateTracesProcessor(context.Background(), ProcessorCreateSettings{}, &defaultCfg, nil)
+	assert.NoError(t, err)
+
+	assert.EqualValues(t, StabilityLevelBeta, factory.StabilityLevel(config.MetricsDataType))
+	_, err = factory.CreateMetricsProcessor(context.Background(), ProcessorCreateSettings{}, &defaultCfg, nil)
+	assert.NoError(t, err)
+
+	assert.EqualValues(t, StabilityLevelUnmaintained, factory.StabilityLevel(config.LogsDataType))
+	_, err = factory.CreateLogsProcessor(context.Background(), ProcessorCreateSettings{}, &defaultCfg, nil)
+	assert.NoError(t, err)
+}
+
 func createTracesProcessor(context.Context, ProcessorCreateSettings, config.Processor, consumer.Traces) (TracesProcessor, error) {
 	return nil, nil
 }